# models/new_cart_response.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// NewCartResponse cart output
//
// swagger:model NewCartResponse
type NewCartResponse struct {

	// cart Id
	CartID string `json:"cartId,omitempty"`
}

// Validate validates this new cart response
func (m *NewCartResponse) Validate(formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *NewCartResponse) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *NewCartResponse) UnmarshalBinary(b []byte) error {
	var res NewCartResponse
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}