# models/update_quantity.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// UpdateQuantity update quantity
//
// swagger:model UpdateQuantity
type UpdateQuantity struct {

	// quantity
	Quantity float64 `json:"quantity,omitempty"`
}

// Validate validates this update quantity
func (m *UpdateQuantity) Validate(formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *UpdateQuantity) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *UpdateQuantity) UnmarshalBinary(b []byte) error {
	var res UpdateQuantity
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}